Convert everything to be compatible with the latest raspberry pi
Extended the install script.
Made bluetooth work with the application
Fixed python module to included the correct data
