+++ pair/config.py
