+++ pair/logger.py
@@ -20,7 +20,7 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-import bluetooth.config as config
+import pair.config as config
 
 LOG_ERROR = "[ERROR]" # type of logging
 LOG_DEBUG = "[DEBUG]" # type of logging
