+++ pair/wifi.py
@@ -1,18 +1,17 @@
-
 # MIT License
-# 
+#
 # Copyright (c) 2019 AP Hogeschool
-# 
+#
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
 # in the Software without restriction, including without limitation the rights
 # to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
-# 
+#
 # The above copyright notice and this permission notice shall be included in all
 # copies or substantial portions of the Software.
-# 
+#
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 # FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -20,23 +19,19 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-import wifi
-import logger
-import bluetoothManager
+import socket
 
-def isConnectedTOWifi():
-    return wifi.ConnectedToTheNetwork()
-
-def establishConnection():
-    bluetoothManager.EstablishConnection()
-
-def start_farm():
-    logger.log("Starting farm", logger.LOG_DEBUG)
+import pair.config as config
+import pair.logger as logger
 
 
-if __name__ == "__main__":
-    if isConnectedTOWifi():
-        logger.log("Connected to the wifi")
-        start_farm()
-    establishConnection()
-    start_farm()
+def ConnectedToTheNetwork():
+    try:
+        socket.setdefaulttimeout(config.NETWORK_TIMEOUT)
+        socket.socket(socket.AF_INET, socket.SOCK_STREAM).connect(
+            (config.NETWORK_CHECK, config.NETWORK_PORT))
+        return True  # TODO: this should change once deployed
+    except socket.error as ex:
+        logger.log("No connection is present to the internet. {}".format(ex),
+                   logger.LOG_ERROR)
+        return False
